web_library/requirements.txt
Django==4.2.5
django-environ==0.11.2
django-ranged-response==0.2.0
django-recaptcha==3.0.0
django-rest-swagger==2.2.0
django-simple-captcha==0.5.20
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
openapi-codec==1.3.2
Pillow==10.0.1
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg2-binary==2.9.7
pytest==7.4.2
pytest-django==4.5.2
python-dotenv==1.0.0
pytils==0.4.1
requests==2.31.0
sqlparse==0.4.4